requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pyyaml
feedparser
jinja2
